make every MCP tool a wrapper over a shared core, not a second implementation
