+++ apps/web/src/routes/api/v1/runs/$runId.ts
@@ -1,18 +1,11 @@
 /**
  * GET /api/v1/runs/:runId — one answer, in full.
- *
- * `answer.text` is the normalized extraction, never the provider's own payload:
- * that shape belongs to the provider, and exposing it would quietly make it
- * part of this API's contract.
  */
 import { createFileRoute } from "@tanstack/react-router";
-import { db } from "@workspace/lib/db/db";
-import { citations, promptRuns } from "@workspace/lib/db/schema";
-import { extractTextContent } from "@workspace/lib/text-extraction";
-import { asc, eq } from "drizzle-orm";
 import { z } from "zod";
 import { ApiError, createApiHandler, withMethodGuard } from "@/lib/api/handler";
 import { isBrandInScope } from "@/lib/api/scope";
+import { findRunDetail } from "@/server/runs-core";
 
 export const Route = createFileRoute("/api/v1/runs/$runId")({
 	server: {
@@ -21,42 +14,12 @@ export const Route = createFileRoute("/api/v1/runs/$runId")({
 				params: z.object({ runId: z.guid("Invalid run ID format") }),
 				scopes: ["runs:read"],
 				handle: async ({ params, auth }) => {
-					const { runId } = params;
-					const [run] = await db.select().from(promptRuns).where(eq(promptRuns.id, runId)).limit(1);
+					const run = await findRunDetail(params.runId);
+					// A run in another tenant reads exactly as one that isn't there.
 					if (!run || !(await isBrandInScope(auth, run.brandId))) {
-						throw new ApiError(404, "Not Found", `Run with ID '${runId}' not found`);
+						throw new ApiError(404, "Not Found", `Run with ID '${params.runId}' not found`);
 					}
-
-					const cited = await db
-						.select({
-							url: citations.url,
-							domain: citations.domain,
-							title: citations.title,
-							citationIndex: citations.citationIndex,
-						})
-						.from(citations)
-						.where(eq(citations.promptRunId, runId))
-						.orderBy(asc(citations.citationIndex));
-
-					// Older rows predate the provider column; the model name is the
-					// extractor's other accepted key, so it is the right fallback.
-					const text = extractTextContent(run.rawOutput, run.provider ?? run.model);
-
-					return {
-						id: run.id,
-						promptId: run.promptId,
-						brandId: run.brandId,
-						model: run.model,
-						provider: run.provider,
-						webSearchEnabled: run.webSearchEnabled,
-						brandMentioned: run.brandMentioned,
-						competitorsMentioned: run.competitorsMentioned,
-						webQueries: run.webQueries,
-						citationCount: cited.length,
-						createdAt: run.createdAt,
-						answer: { text: text || null },
-						citations: cited,
-					};
+					return run;
 				},
 			}),
 		}),
